requirements.txt
openai
chromadb
google-generativeai
streamlit
docling
chonkie[all]